src/requirements.txt
scrapy
pymongo
mysql-connector
PySocks
